Lessons 3 level_1 removed public and refactor all class
