+++ src/main/java/students/student_ilja_morozovs/lesson_3/homework/level_1/Task_1/Robot.java
@@ -1,0 +1,8 @@
+package students.student_ilja_morozovs.lesson_3.homework.level_1.Task_1;
+
+class Robot {
+
+    public void sayHello() {
+        System.out.println("Hello!");
+    }
+}
