+++ src/main/java/students/student_ilja_morozovs/lesson_3/homework/level_1/Task_3/Robot.java
@@ -1,0 +1,14 @@
+package students.student_ilja_morozovs.lesson_3.homework.level_1.Task_3;
+
+public class Robot {
+    public Robot() {
+    }
+
+    public void sayHello() {
+        System.out.println("Hello!");
+    }
+
+    public void sayYourName() {
+        System.out.println("My name is ROBO");
+    }
+}
