+++ src/main/java/students/student_ilja_morozovs/lesson_3/homework/level_1/Task_3/RobotDemo.java
@@ -1,0 +1,10 @@
+package students.student_ilja_morozovs.lesson_3.homework.level_1.Task_3;
+
+
+class RobotDemo {
+    public static void main(String[] args) {
+        Robot verter;
+        verter = new Robot();
+
+    }
+}
