+++ src/main/java/students/student_ilja_morozovs/lesson_3/homework/level_1/Task_4/RobotDemo.java
@@ -1,0 +1,9 @@
+package students.student_ilja_morozovs.lesson_3.homework.level_1.Task_4;
+
+
+class RobotDemo {
+    public static void main(String[] args) {
+        Robot verter = new Robot();
+        verter.sayHello();
+    }
+}
