+++ src/main/java/students/student_ilja_morozovs/lesson_3/homework/level_1/Task_6/RobotDemo.java
@@ -1,0 +1,15 @@
+package students.student_ilja_morozovs.lesson_3.homework.level_1.Task_6;
+
+
+class RobotDemo {
+    public static void main(String[] args) {
+        Robot verter;
+        verter = new Robot();
+        verter.sayYourName();
+        verter.sayHello();
+        System.out.println("________________");
+        Robot verter1 = new Robot();
+        verter1.sayHello();
+        verter1.sayYourName();
+    }
+}
