--- src/main/java/students/student_ilja_morozovs/lessons_3/homework/level_1/Task_1/Robot.java
@@ -1,8 +1,0 @@
-package students.student_ilja_morozovs.lessons_3.homework.level_1.Task_1;
-
-public class Robot {
-
-    public void sayHello() {
-        System.out.println("Hello!");
-    }
-}
