--- src/main/java/students/student_ilja_morozovs/lessons_3/homework/level_1/Task_2/Robot1.java
@@ -1,12 +1,0 @@
-package students.student_ilja_morozovs.lessons_3.homework.level_1.Task_2;
-
-public class Robot1 {
-    public Robot1() {
-    }
-    public void sayHello() {
-        System.out.println("Hello!");
-    }
-    public void sayYourName(){
-        System.out.println("My name is ROBO");
-    }
-}
