--- src/main/java/students/student_ilja_morozovs/lessons_3/homework/level_1/Task_3/RobotDemo.java
@@ -1,10 +1,0 @@
-package students.student_ilja_morozovs.lessons_3.homework.level_1.Task_3;
-
-import students.student_ilja_morozovs.lessons_3.homework.level_1.Task_2.Robot1;
-
-public class RobotDemo {
-    public static void main(String[] args) {
-        Robot1 verter;
-        verter=new Robot1();
-    }
-}
