--- src/main/java/students/student_ilja_morozovs/lessons_3/homework/level_1/Task_5/RobotDemo.java
@@ -1,11 +1,0 @@
-package students.student_ilja_morozovs.lessons_3.homework.level_1.Task_5;
-
-import students.student_ilja_morozovs.lessons_3.homework.level_1.Task_2.Robot1;
-
-public class RobotDemo {
-    public static void main(String[] args) {
-        Robot1 verter;
-        verter = new Robot1();
-        verter.sayYourName();
-    }
-}
