--- src/main/java/students/student_ilja_morozovs/lessons_3/homework/level_1/Task_6/RobotDemo.java
@@ -1,16 +1,0 @@
-package students.student_ilja_morozovs.lessons_3.homework.level_1.Task_6;
-
-import students.student_ilja_morozovs.lessons_3.homework.level_1.Task_2.Robot1;
-
-public class RobotDemo {
-    public static void main(String[] args) {
-        Robot1 verter;
-        verter = new Robot1();
-        verter.sayYourName();
-        verter.sayHello();
-        System.out.println("________________");
-        Robot1 verter1=new Robot1();
-        verter1.sayHello();
-        verter1.sayYourName();
-    }
-}
